backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.7